+++ lib/instrumentation.ex
@@ -8,8 +8,11 @@ defmodule OpentelemetryAbsinthe.Instrumentation do
   (you can still call `OpentelemetryAbsinthe.Instrumentation.setup()` in your application startup
   code, it just won't do anything.)
   """
+  alias Absinthe.Blueprint
 
   require OpenTelemetry.Tracer, as: Tracer
+  require OpenTelemetry.SemanticConventions.Trace, as: Conventions
+  require Logger
   require Record
 
   @span_ctx_fields Record.extract(:span_ctx,
@@ -18,13 +21,20 @@ defmodule OpentelemetryAbsinthe.Instrumentation do
 
   Record.defrecord(:span_ctx, @span_ctx_fields)
 
+  @default_operation_span "GraphQL Operation"
+  @graphql_document_attr Atom.to_string(Conventions.graphql_document())
+  @graphql_operation_name_attr Atom.to_string(Conventions.graphql_operation_name())
+  @graphql_operation_type_attr Atom.to_string(Conventions.graphql_operation_type())
+
   @default_config [
-    span_name: "absinthe graphql resolution",
+    span_name: :dynamic,
     trace_request_query: true,
-    trace_request_variables: true,
-    trace_response_result: true,
-    trace_response_errors: true,
-    trace_request_selections: true
+    trace_request_name: true,
+    trace_request_type: true,
+    trace_request_variables: false,
+    trace_request_selections: true,
+    trace_response_result: false,
+    trace_response_errors: false
   ]
 
   def setup(instrumentation_opts \\ []) do
@@ -33,6 +43,10 @@ defmodule OpentelemetryAbsinthe.Instrumentation do
       |> Keyword.merge(Application.get_env(:opentelemetry_absinthe, :trace_options, []))
       |> Keyword.merge(instrumentation_opts)
       |> Enum.into(%{})
+
+    if is_binary(config.span_name) do
+      Logger.warn("The opentelemetry_absinthe span_name option is deprecated and will be removed in the future")
+    end
 
     :telemetry.attach(
       {__MODULE__, :operation_start},
@@ -55,28 +69,44 @@ defmodule OpentelemetryAbsinthe.Instrumentation do
   end
 
   def handle_operation_start(_event_name, _measurements, metadata, config) do
-    params = metadata |> Map.get(:options, []) |> Keyword.get(:params, %{})
+    document = metadata.blueprint.input
+    variables = metadata |> Map.get(:options, []) |> Keyword.get(:variables, %{})
 
     attributes =
       []
       |> put_if(
         config.trace_request_variables,
-        {"graphql.request.variables", Jason.encode!(params["variables"])}
+        {"graphql.request.variables", Jason.encode!(variables)}
       )
-      |> put_if(config.trace_request_query, {"graphql.request.query", params["query"]})
+      |> put_if(config.trace_request_query, {@graphql_document_attr, document})
 
     save_parent_ctx()
 
-    new_ctx = Tracer.start_span(config.span_name, %{attributes: attributes})
+    span_name = span_name(nil, nil, config.span_name)
+    new_ctx = Tracer.start_span(span_name, %{attributes: attributes})
 
     Tracer.set_current_span(new_ctx)
   end
 
   def handle_operation_stop(_event_name, _measurements, data, config) do
+    operation_type = get_operation_type(data)
+    operation_name = get_operation_name(data)
+
+    span_name = span_name(operation_type, operation_name, config.span_name)
+    Tracer.update_name(span_name)
+
     errors = data.blueprint.result[:errors]
 
     result_attributes =
       []
+      |> put_if(
+        config.trace_request_type,
+        {@graphql_operation_type_attr, operation_type}
+      )
+      |> put_if(
+        config.trace_request_name,
+        {@graphql_operation_name_attr, operation_name}
+      )
       |> put_if(
         config.trace_response_result,
         {"graphql.response.result", Jason.encode!(data.blueprint.result)}
@@ -99,18 +129,31 @@ defmodule OpentelemetryAbsinthe.Instrumentation do
     :ok
   end
 
-  defp get_graphql_selections(data) do
-    case data do
-      %{blueprint: %{operations: [_ | _] = operations}} ->
-        operations
-        |> Enum.flat_map(& &1.selections)
-        |> Enum.map(& &1.name)
-        |> Enum.uniq()
-
-      _ ->
-        []
-    end
+  defp get_graphql_selections(%{blueprint: %Blueprint{} = blueprint}) do
+    blueprint
+    |> Blueprint.current_operation()
+    |> Kernel.||(%{})
+    |> Map.get(:selections, [])
+    |> Enum.map(& &1.name)
+    |> Enum.uniq()
   end
+
+  def default_config do
+    @default_config
+  end
+
+  defp get_operation_type(%{blueprint: %Blueprint{} = blueprint}) do
+    blueprint |> Absinthe.Blueprint.current_operation() |> Kernel.||(%{}) |> Map.get(:type)
+  end
+
+  defp get_operation_name(%{blueprint: %Blueprint{} = blueprint}) do
+    blueprint |> Absinthe.Blueprint.current_operation() |> Kernel.||(%{}) |> Map.get(:name)
+  end
+
+  defp span_name(_, _, name) when is_binary(name), do: name
+  defp span_name(nil, _, _), do: @default_operation_span
+  defp span_name(op_type, nil, _), do: Atom.to_string(op_type)
+  defp span_name(op_type, op_name, _), do: "#{Atom.to_string(op_type)} #{op_name}"
 
   # Surprisingly, that doesn't seem to by anything in the stdlib to conditionally
   # put stuff in a list / keyword list.